fix: Controller에서 의뢰인에 대한 사건 조회 URL 수정

+++ src/main/java/com/avg/lawsuitmanagement/client/repository/ClientMapperRepository.java
@@ -2,11 +2,8 @@ package com.avg.lawsuitmanagement.client.repository;
 
 import com.avg.lawsuitmanagement.client.dto.ClientDto;
 import com.avg.lawsuitmanagement.client.repository.param.InsertClientParam;
-import com.avg.lawsuitmanagement.client.repository.param.SelectClientLawsuitListParam;
 import com.avg.lawsuitmanagement.client.repository.param.UpdateClientInfoParam;
 import com.avg.lawsuitmanagement.client.repository.param.UpdateClientMemberIdParam;
-import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitDto;
-import java.util.HashMap;
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 
@@ -20,6 +17,6 @@ public interface ClientMapperRepository {
     void updateClientInfo(UpdateClientInfoParam param);
     void deleteClientInfo(long clientId);
     List<ClientDto> selectClientList();
-    long getLawsuitCountByClientId(long clientId);
+    int getLawsuitCountByClientId(long clientId);
     List<ClientDto> selectClientListById(List<Long> clientIdList);
 }
